Added an inversion test for N = 7 and fixed leftLaInv complex number behaviour

+++ clifford/__init__.py
@@ -256,7 +256,8 @@ def grades_present(objin: 'MultiVector', threshold=0.0000001) -> Set[int]:
 # todo: work out how to let numba use the COO objects directly
 @_numba_utils.njit
 def _numba_val_get_left_gmt_matrix(x, k_list, l_list, m_list, mult_table_vals, ndims):
-    intermed = np.zeros((ndims, ndims))
+    # TODO: consider `dtype=result_type(x.dtype, mult_table_vals.dtype)`
+    intermed = np.zeros((ndims, ndims), dtype=x.dtype)
     test_ind = 0
     for k in k_list:
         j = l_list[test_ind]
